+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Fixed
 
+- Fix false out-of-band authentication failures on update and clone ([740])
+
+[740]: https://github.com/openlawlibrary/taf/pull/740
+
 
 ## [0.38.3]