+++ CHANGELOG.md
@@ -14,10 +14,13 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Fixed
 
+
 - Fix commits on branch and not other when there is no local merge branch ([744])
+- Raise clear error when default branch not found in local repo [(742)]
 - Fix false out-of-band authentication failures on update and clone ([740])
 
 [744]: https://github.com/openlawlibrary/taf/pull/744
+[742]: https://github.com/openlawlibrary/taf/pull/742
 [740]: https://github.com/openlawlibrary/taf/pull/740